backend/requirements.txt
aiofiles==0.6.0
alembic==1.5.8
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.2.0
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
emails==0.6
fastapi==0.66.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
inflect==5.3.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
premailer==3.7.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.16.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0